#sdy add a pass to apply sharding constraints that dictate the sharding of their input.
Copies the sharding of a `ShardingConstraintOp` to its input if it satisfies all of the following:

- The input doesn't have an existing sharding.
- The input is either only used by the `ShardingConstraintOp` or the latter doesn't have any uses (dangling) and the input doesn't have any other users of type `ShardingConstraintOp`.

Which indicates that the `ShardingConstraintOp` dictates the sharding of its input.

Note that the sharding of a `ShardingConstraintOp` will propagate to its input or users during propagation, but since the closed property of a dimension doesn't propagate, it's important to copy the sharding to respect the constraint in the above cases.

PiperOrigin-RevId: 651717821