#sdy Export pass to make function inputs/outputs evenly sharded.
After propagation, function inputs/outputs may require padding. We can't allow this. GSPMD makes the tensor fully replicated if one dimension requires padding, however we can improve on that. We find the best sub-sharding per dimension by taking the largest dimension sharding prefix that would only require basic communication - a dynamic-slice for the input tensor, or a basic all-gather on the output tensor.

Note that as part of the work to add explicit reshard ops, we need to make it add reshards for these inputs/outputs, so the explicit reshard pass should come after this pass.

PiperOrigin-RevId: 660367858