#sdy add `sdy::NamedComputationOp` in/out shardings and make it adhere to `ShardableDataFlowOpInterface`, and add op result data flow op interface methods.
This means Shardy can now add/sink dataflow edges to and propagate through `sdy::NamedComputation` ops.

The added methods match the equivalent `ShardableDataFlowOpInterface` on block arguments. This makes the interface more complete, polishes up the utility functions, but will also be important when we make `ManualComputationOp` adhere to the `ShardableDataFlowOpInterface` as it will be overriding these new methods (it needs to set/get the shardings from the registered `out_shardings` op argument.

PiperOrigin-RevId: 679155557