#sdy fix bug in reshape sharding rule
Before this change, we would greedily add a unique factor for both input and output as long as there is a mismatch in accumulated dim sizes, but then we might not allow them to match again. This results in missing common factors and also failing `assert(inDim < inRank && outDim < outRank)` since one tensor reached the end before the other.

PiperOrigin-RevId: 687668725