Shardy. Sort the factors before resolving conflicts across factors in `AggressiveFactorPropagation::propagateFactorShardings`.
We may resolve real conflicts across factors in the next cls. This one collects the map from "factor index" to "the size of the largest source tensor." We may prefer the factor with larger size of source tensor when we resolve real conflicts.

This cl does not resolve real conflicts and should not impact the propagation result.

We define the "source tensor" as follows. We collect new axes along a factor. If the new axes are not empty, we can classify tensors in three buckets.
1. The factor is missing in this tensor.
2. We cannot propagate the new sharding to the tensor. The tensor must have "better" sharding axes and is the "source" of the new axes.
3. We can propagate the new sharding to the tensor, i.e., this tensor is a "target".

PiperOrigin-RevId: 689161756