Define close-shardings pass.
The pass closes all shardings, including func input and output ones, and drops replicated axes as well.

The pass is meant to be on the export pipeline, as the first pass of it. The pass is not enabled yet, but only defined.

The pass of inserting explicit reshards depend on this pass as it assumes all shardings are closed and replicated axes are dropped.

PiperOrigin-RevId: 695422467