Define drop-sharding-rules pass.
The pass drops sharding rules from all registered ops.

The pass is meant to be on the export pipeline. The pass is not enabled yet, but only defined.

This pass, for performance reasons only, needs to be after the pass of inserting explicit reshards so that we do not recreate sharding rules during inserting explicit reshards. The sharding rules are created during propagation.

PiperOrigin-RevId: 695706637