Use majority vote heuristic to find common axes for factors, for single full axes.
It keeps using Greatest Common Prefix heuristic for cases where some factors have subaxes or multiple axes, which will be handled subsequently.

The heuristic is, at each iteration, to find a {factor, axis} pair with the largest count from the list of 'valid' pairs (initially all pairs with non-zero count are valid) and drop all the pairs with the picked factor or with an axis that overlaps with the picked axis from the list. Iterate until the list is empty.

PiperOrigin-RevId: 695742802