#sdy Add debugging tool to show the origin shardings on the MLIR module.
This works by creating initial source nodes into an `mlir::Value` map, and then every propagation step where we update another values sharding we add a new entry using the entry of the source node we put in initially. This happens iteratively for each propagation step.

To determine whether something has been updated, we just look at the old and new sharding of the `Value`, and if they are different, look for what axis is new, and look up the axis in a locally built `axis-to-value` map.

Note that the source sharding works by building up edge shardings, which we will be needing when building the Model Explorer graph.

Only thing unsupported in this CL are `ManualComputationOp`s with manual axes and ops with multiple results.

1. The issue for `ManualComputation` is that we call `getSharding(value)` to get the new sharding of a `Value`. However, this doesn't account for the `sdy::ReturnOp`:`ManualComputationOp` barrier, where we actually call `getOutShardingWithoutManualAxes` during propagation. Instead of plumbing through a call-back to get the sharding (like we have `setSharding`), when we make `ManualComputationOp` adhere to the `ShardableDataFlowOpInterface`, then we won't need this anymore.

2. The issue with multiple results is I need to make `saveShardingOriginsOnModule` handle it when the op is a generic `Operation*`. Will do this in a follow-up since the PR is pretty big.

PiperOrigin-RevId: 699955947