Tie break factor-axes pairs on sharded tensor size when the counts are the same, favor the larger one.
In case the factor indices are the same, or the tensor sizes are the same, then tie-break on the factor-axes pair with larger axes to further-shard.

Use the source tensor size, and when there are multiple source tensor sizes, use the largest one. This is to keep shardings originated from bigger tensors so to minimize resharding cost.

PiperOrigin-RevId: 700278772