Reserve backing data for axis list ref tests.
This is to keep the original vectors as they are referenced from created axes. It should reserve enough to accommodate a single unit-test.

For example, if the contents of the vector reaches its capacity, it may create a new array with a larger capacity and copy the contents, which would invalidate the references.

It also fixes an msan failure as filed on b/381409526.

PiperOrigin-RevId: 701287521