Add a unit test for insert-explicit-reshards pass for stablehlo.reverse.
stablehlo.reverse is a special operation in that it adds collective permute during partitioning. Yet, at least for the non-reversing dimensions, a conflict resolution is still needed. This unit test verifies that explicit reshards are added for conflict resolution.

That said, for reversing dimensions, there possibly is room to handle conflict resolution more efficiently, which is not in the scope of this CL.

PiperOrigin-RevId: 712469397