Do not insert explicit reshards for operation that has factors that are specially handled by the partitioner.
It is possible that adding explicit reshards on a conflict resolution on special factors can lead to programs that performs worse than what the partitioner would produce when they are not added. For example, for sort operation, the sorting dimension is fully replicated before the sort operation, and adding an explicit reshard to partition the sorting factor could lead to an unnecessary all-gather as the sort operation will fully-replicate it.

See go/stablehlo-ops-sharding for more through evaluation of stablehlo operations with special dimensions.

PiperOrigin-RevId: 714397498