Shortcut to find common axes when the operation is unary and only one tensor is incompatible individually.
It covers BitcastConvert and the cases that one resharding is sufficient for operations with factors that need replication, such as Sort, and Cholesky.

Individual incompatibly, at the moment, happens only in the presence of fully replicated factors, and a tensor is individually incompatible if its factor that need replication is sharded.

PiperOrigin-RevId: 718355522