Reshard only one tensor when possible for cheap operations. part 1.
Hence, exclude the operations that need replication.

In this part, reshard the only tensor that makes others' sharding compatible without. The second part will pick the better one in case there are two such tensors.

The following is a sufficient and necessary condition for a tensor T to leave the others compatible:
1. All factors are sharded at most two different ways across tensors. Call the ones that are sharded the same as compatible factors and that are sharded differently as incompatible factors.
2. Tensor T contains all incompatible factors and its sharding is unique (across tensors) for each of those incompatible factors.
3. The factor shardigns on other tensors do not overlap.

PiperOrigin-RevId: 718830988