#sdy Initial stage of converting any `sdy.reshard` to collectives.
The pass converts `sdy.reshard` to a sequence of `sdy.all_gather` and `sdy.all_slice` in a naive way, by trying to all-slice what we can with available axes, then all-gathering axes that aren't where they need to be in the result sharding, finally all-slicing some of the gathered axes as needed.

In follow up changes, we will introduce `sdy.all_to_all` and `sdy.collective_permute`, and improve this pass to give the optimal sequence of collectives.

PiperOrigin-RevId: 719641957