#sdy Introduce `sdy.collective_permute`
An axis based collective permute can replace axes across dimensions with the same or other axes, and in a different order, as long as the sharded size in each dimension stays the same. This translates into a single HLO collective permute.

For example:
```
sdy.mesh @mesh = <["a"=2, "b"=2, "c"=4, "d"=2, "e"=2, "f"=2]>
%1 = stablehlo.tanh(%0) {sdy.sharding = #sdy.sharding_per_value<[<@mesh, [{"a", "c"}, {"f"}, {"d", "e"}]>]>} : tensor<8x8x8xf32>
%2 = sdy.collective_permute %1 out_sharding=<@mesh, [{"c":(1)2, "b", "f"}, {"a"}, {"e", "d"}]> : tensor<8x8x8xf32>
```

PiperOrigin-RevId: 720572879