#sdy when inserting all-to-alls, first try to insert all-to-alls that move axes to the right position at the target dimension.
This is important to make sure multiple all-to-alls from different source dimensions to the same target dimension are inserted in the right order (to avoid the need for a collective permute).

We then lift this constraint for additional all-to-alls to be inserted. This is only needed for cases where axes are not divisible (not all a power of 2).

PiperOrigin-RevId: 727871138