Do not set the sharding of the sharding group op during import, and emit a warning instead of error if the group has a conflicting initial sharding.
What we should do instead is let propagation resolve what sharding to use. Note this will be a bit arbitrary based on propagation order. We should still at least emit a warning to the user since this is a bit of a weird use of sharding group:

```mlir
/tmp/bug.mlir:9:3: warning: Inconsistent shardings prior to propagation for ShardingGroupOps with canonicalized groupId: 0 due to the operand %1 = sdy.sharding_constraint %arg0 <@mesh, [{"a"}, {?}]> : tensor<8x8xf32>. Inserting an open sharding constraint to all constrained  values
  sdy.sharding_group %1 group_id=1183 : tensor<8x8xf32>
  ^
/tmp/bug.mlir:9:3: note: see current operation: sdy.sharding_group %1 group_id=0 : tensor<8x8xf32>
```

This also matches the behavior in GSPMD.

PiperOrigin-RevId: 727888959