Avoid the second basic propagation iteration to confirm convergence in production builds.
We know that a single iteration is enough to converge as we add affected ops back to the worklist. Removing the extra iteration for convergence check reduces the pass time by ~33%. This would also allow op priority propagation to have more iterations without the overhead of the extra convergence check iteration.

In debug builds, we won't fail if failed to converge, but issue a warning instead.

In addition, remove strided view warning as we no longer need it to know there are cases that would benefit from support.

PiperOrigin-RevId: 732412330