#sdy Apply a sharding constraint when the input is a `DataFlowEdgeOp`.
We previously didn't do this to match GSPMD, as they only set the sharding of the corresponding get-tuple-element, not the data flow op itself (while, case, opt_barrier).

However, when the `DataFlowEdgeOp` has other uses, those wouldn't get the sharding of the sharding constraint, which is unwanted behavior and doesn't match GSPMD.

To resolve this while still not directly setting the sharding of the `DataFlowEdgeOp`, which affects all targets of the edge, we replace all uses of the op with the sharding constraint.

Note that if the constraint is a `ShardingConstraintOp`, we can simply move it next to the `DataFlowEdgeOp`, but if it's a `ManualComputationOp`, we need to create a `ShardingConstraintOp` to put next to the `DataFlowEdgeOp`.

PiperOrigin-RevId: 733741860