#sdy remove `Elementwise` traits from shardy ops, make it part of the `isElementwise` func.
These ops aren't really element-wise, and interact incorrectly with StableHLO passes. We only marked this as element-wise so that in op priority propagation pass, we propagate them at the same time as regular element-wise ops (like in StableHLO). It's also used by the constant splitting import pass.

So instead, just update the `isElementwise` function that is used by Shardy.

PiperOrigin-RevId: 736888665