#sdy Insert `sdy.all_reduce` ops in `-sdy-insert-explicit-reshards` pass.
If an op has a sharded reduction factor (e.g., contracting dim of dot) after reshards are inserted, we should insert a `sdy.all_reduce` with the axes that shard all reduction factors.

In the future, when we have support for unreduced axes in TensorShardingAttr, we can have this pass insert a reshard from unreduced axes to replicated axes, and insert the all-reduce later on.

PiperOrigin-RevId: 737966111