[shardy] Fix type mismatch between header and implementation for sdyOpShardingRuleAttrGet.
This was breaking the JAX build because it meant that the function signature mismatches and the definition was not considered extern "C".

To make sure the compiler catches this kind of error in the future, wrap all of the exports in an extern "C". This will make at least some compilers complain about the function signature mismatch, since C functions cannot be overloaded.

PiperOrigin-RevId: 743553902