#sdy Add a sharding rule to `stablehlo.rng_bit_generator` that forces that initial and output state to be replicated and block any propagation between them.
In addition, add a special sharding rule to `@X64Combine`, if any of its users is a `stablehlo.rng_bit_generator`, that blocks propagation along all its factors. This is to ensure that in such a case, the sharding of its inputs don't propagate to the result, which would ensure the initial state of the `stablehlo.rng_bit_generator` is replicated and no reshard is inserted on the result of the `@X64Combine`.

Finally, add `@InspectSharding` custom call to the registry, it was missing before.

PiperOrigin-RevId: 753766698