Do not restrict on resharding only one tensor for unary operations when some dimensions have multiple factors.
Which indicates splits and merges of factors, in which case, the resharding decisions should be more flexible.

After this change those cases use majority vote to decide on resharding, however, it is not a final decision yet. The idea is to count factor axis candidates yet invalid if on some dimensions with multiple factors they cause strided views.

PiperOrigin-RevId: 755769174