Avoid strided views for common axes for reshapes.
Also for other ops with dimensions with multiple factors, such as, convolution.

This change introduces the concept of factor dependencies: A factor f depends on factor g if two factors appear together in any tensor dimension, and g appears before f. A factor is non-full if its sharding size is smaller than the size of the factor. This list is used to determine if a factor-axes candidate is valid yet since a factor should not be sharded until it does have zero dependencies, that is, all factors that appear, in any tensor dimension, before the factor needs to be fully-sharded, otherwise it would introduce a strided view which is not supported yet.

PiperOrigin-RevId: 757818248