Add an example that propagation should propagate through reduction factor on a dot when the rhs is considerably large.
In this example, dot performs on [4][1024] x [1024][1024] = [4][1024], and for its sharding rule, [i][k],[k][j] = [i][j] it is sharded as [][],[x][] = [x][], before the propagation.

It currently gets [x][],[x][] = [x][] after the propagation, which is either because it propagated from result to lhs, or it is propagated from the input of the defining op (which is broadcast in this example) of lhs to its output. GSPMD partitioning adds an all-gather on rhs in this example, which is costly since rhs is a large tensor.

It instead should propagate from rhs to lhs, so it gets sharded as [][x],[x][] = [x][]. GSPMD in this case, would perform an all-reduce and a dynamic slice on the output.

PiperOrigin-RevId: 758208626