Introduce unreduced axes in TensorShardingAttr.
A sharding can have unreduced axes (specified by `unreduced_axes`), meaning the tensor is unreduced along these axes. Applying an all-reduce on the tensor along the unreduced axes will make the tensor replicated along those axes. However, a tensor with unreduced axes doesn't have to be all-reduced immediately, it can remain unreduced when passed to linear operations like `stablehlo.add` (as long as both lhs and rhs are unreduced) and all-reduced afterwards. We assume the reduction type is sum, other reductions may be supported in the future.

Note that the change is both bwd and fwd compatible, due to the versioning added for TensorShardingAttr in bytecode.td.

PiperOrigin-RevId: 758243917