#sdy Insert an `sdy.all-reduce` on the use of a tensor that has unreduced axes or sub-axes that aren't unreduced in the result sharding of the user.
In addition, don't insert an all-reduce for reduction factors if the result of the op is unreduced on the axes sharding that factor.

We take the set differences between unreduced axes of input and unreduced axes of output, to get the all-reduce axes.

We add this logic in both `-sdy-insert-explicit-reshards` and `-sdy-temp-explicit-reshards-for-optimizations` so it will be applied regardless of whether explicit collectives are enabled.

PiperOrigin-RevId: 764298966