#sdy Insert an `sdy.all_reduce` on the use of a tensor that has unreduced axes or sub-axes that aren't unreduced in the result sharding of the user.
In addition, don't insert an all-reduce for reduction factors if the result of the op is unreduced on the axes sharding that factor.

We take the set differences between unreduced axes of input and unreduced axes of output, to get the all-reduce axes.

NOTE: we don't verify correctness of an operation that takes an unreduced input and produces an unreduced output, we simply respect the existing shardings and match cases where an unreduced axis becomes sharded/replicated. It's the responsibility of the user/frontend to only push an unreduced axis through an operation (e.g., add) when it's semantically correct.

We add this logic in both `-sdy-insert-explicit-reshards` and `-sdy-temp-explicit-reshards-for-optimizations` so it will be applied regardless of whether explicit collectives are enabled.

PiperOrigin-RevId: 767735351