#sdy Bring back blocking sideways propagation.
Blocking sideways propagation makes sense in practice because for example, in `c = add(a, b)`, if `a` is sharded and `c` is replicated, it doesn't make sense to propagate to `b` without more information.

We originally didn't, then we blocked it since it improved the performance of propagation on a production model, then we rolled it back again since there was a change in the XLA partitioner that fixed the use of it in the production model, and we still haven't validated it across different production models. But now we have and see it has improved another production model. So we are bringing it back.

Reverts 651afd0c31ae3ccf7b80c6a543c48da8578357f1

PiperOrigin-RevId: 789884506