Support factor types in sharding rules for custom partitioning.
A factor in sharding rule is like an index label in Einsum notation. A factor usually corresponds to a dimension; the type of the dimension, such as passthrough/elementwise or reduction, is indicated by the type of the factor. By default, a factor is of passthrough type. Users can overwrite this default factor type through keyword arguments.

PiperOrigin-RevId: 793737548