Insert all reduces for unreduced axes also when the source unreduced axes is a strict subset of target unreduced axes.
So that the source and target have the same unreduced axes, and all reduce axes is empty.

In which case, allReducedAxes is still empty, but the source and target unreduced axes mismatch. After this change, the allreduces for unreduced axes guarantees that source and target shardings have the same unreduced axes during reshard on func results and data flow ops.

PiperOrigin-RevId: 805407301