[mpmd] Check SplitFragmentType and remat flag when verifying stage merging
SplitBwdFragmentsPass introduces fragments which previously would cause
VerifyStageMergingPass to fail. This change updates the verification logic to
also differentiate fragments based on SplitFragmentType and remat status,
ensuring that split backward and remat fragments are not incorrectly
identified as duplicates.

Also rename FragmentData to FragmentSignature to better differentiate it from
FragmentInfo.

PiperOrigin-RevId: 806425882