[mpmd] Update call_count and transpose_count in merged remat fragments.
We set call_count in merged_fragment with its value in backward_fragment.

The transpose count of merged remat fragments is determined by the maximum of its multiple origin transpose counts. The rationale of this decision is as follows: A merged fragment can have multiple origin transpose counts, e.g., when merging a rematerialized forward fragment A (transpose_count=0) with a backward fragment B (transpose_count=1). Taking the max transpose count ensures that we are reflecting the highest order of differentiation the fragment is involved in: in this case, the merged fragment A+B is still fundamentally part of the first derivative calculation.

PiperOrigin-RevId: 811858396