Avoid reshards on named computation inputs optionally.
By default, it does not avoid reshards, hence this change is no-op for shardy.

When dedupping functions fully, which happens during StableHLO
export, the input sharding of the function may end up being
different than the input sharding of the named computation during
inserting explicit reshards. Hence, it should avoid inserting
explicit reshards to resolve mismatch on the sharding of the
argument to the named computation and its input sharding.

PiperOrigin-RevId: 823475320